+++ flipper/hotspot-arcade/ha_proto.h
@@ -16,7 +16,7 @@
 #define HA_FW_MAGIC_1 0x41
 #define HA_FW_MAGIC_2 0x52
 #define HA_FW_MAGIC_3 0x43
-#define HA_FW_VERSION 17 // v17: Chess game added
+#define HA_FW_VERSION 18 // v18: Spyfall game added
 
 // Flipper -> ESP
 #define HA_MSG_CLEAR_FILES   0x10
@@ -61,6 +61,7 @@
 #define HA_GAME_SPECTRUM 13 // wavelength-style spectrum guessing (party)
 #define HA_GAME_KMK      14 // kiss marry kill (party)
 #define HA_GAME_CHESS    15 // chess (1v1, full FIDE rules)
+#define HA_GAME_SPYFALL  19 // spyfall (party, one player doesn't know the location)
 
 static inline uint8_t ha_crc8_upd(uint8_t crc, uint8_t b) {
     crc ^= b;